changed discussions into a model, modularized it
