+++ server/src/models/index.ts
@@ -2,5 +2,6 @@ import { Book } from "./book";
 import { Group } from "./group";
 import { User } from "./user";
 import { Review } from "./review";
+import { Discussion } from "./discussion";
 
-export { Book, Group, User, Review }
+export { Book, Group, User, Review, Discussion }
